addLink: bugfix and QoL improvements
All changes in pdf.PdfFileWriter.addLink:
Added documentation to make use a bit clearer.
Fixed an issue where /P was mapped to a repr'd IndirectObject instead of
a proper indirect reference.
Added support for a border-styling parameter.
Allowed rect to be a string (as before), or a RectangleObject, or a
list, to make use more intuitive.