PdfObject: add function to check whether an object refers to the null object
The only objects that can refer to the null object are the null object itself
and indirect references to nonexisten objects. In principle, there could be
an indirect reference to an existing object with value null. Catching this one
would require resolving the indirect reference, which is too much effort.