Handles exception thrown by data.pdf.getObject
Trying to copying a "corrupted" PDF file into a new file was throwing the Exception utils.PdfReadError("Could not find object."), and the application was unable to make the actual copy. This corrupted PDF, however, would work just fine in PDFReader, and other PDF viewers. 

My guess on the take on the issue is that my file contains a reference to a resource/object that was removed and no longer exists. If the object no longer exist, I am sure, it is safe to ignore when copying. 

This is a snippet of the code I was using to copy the file. Although I was modifying the file metadata, the error would still even without modifying the metadata.

```
        original_pdf = PdfFileReader(save_path, strict=False)
        original_pdf_meta_data = get_metadata(original_pdf)
        original_pdf_title = original_pdf_meta_data.get('/Title', '')
        if original_pdf_title != metadata.get('/Title'):
            logger.debug("Creating a new PDF file with updated metadata for "
                         "publication {}".format(pub_busid))
            # only write the new metadata if we are changing it
            output_pdf = PdfFileWriter()
            for page in range(original_pdf.getNumPages()):
                output_pdf.addPage(original_pdf.getPage(page))
            # save updated PDF to temp location
            with open(temp_path, 'wb') as outstream:
                output_pdf.addMetadata(metadata)
                output_pdf.setPageLayout('/SinglePage')
                output_pdf.write(outstream)
                logger.debug('The PDF Metadata has been updated for "{}"'.format(save_path))
```