Fix issue #355.
Borrowed code from ademidun in the comment history and inserted it into the
proper location in the pdf.py module.

Made some changes to the function to make it a method of the class.

It appears to work. I don't have a huge test suite set up to check it.