Glyph encoding for Type1C
The names of the glyphs seem to be "known" with their unicode mappings.
This commit brings such name mappings from two different github projects
to create a common list.
Then the /Differences array is read and parsed into a CMAP compatible
structure so that the translation works.

This removes the hard coded encoding used for Amex and Mastercard Hebrew
documents.