Increase tolerance for garbage after EOF
The electricity bill has some garbage (HTML page) after the EOF marker...
PyPDF allows up to 1K of such garbage so the file is rejected.
Increased it to make it work.
TODO: need a more robust solution, to accept files with more garbage.
One way to do it is to cut that garbage before uploading.