Handle graphics state
Text coordinates are provided in a transformed space,
which is controlled by the 'cm' operator.
In this commit, we read the cm matrix, take only the x and y scale
operands and store them in a stack.
When text operands arrive, we use the y value to translate the Y value.
When q and Q operators we push and pop the graphics state respectively.
For example, a pdf can push a new cm when displaying an image,
and then go back to the previous cm values.
These are all demonstrated in the supergas bill...