extract the correct char widths
When a Tm operator appears, save that matrix in the text state
This is needed, for example in the BezeqPhone bill, where the font size
is always 1 and the text actual size is controlled by the X and Y
multipliers in the text matrix.

The width calculation is integrated into the translation operation
because
both these operations go over the glyph IDs so instead of calcualting
the byte width and iterating again, just do it in the same loop.