Add Table of Contents feature to `pdfcat`.
The TOC is automatically generated based on the -T flag. No custom naming
of the entries is currently supported.
The value is based on the `title` attribute of `DocumentInformation`; if it is
missing, then the original filename is used as a source for
that (e.g.  `Title.pdf` becomes `Title`, and `/a/b/Title.pdf` just `Title`).