Fix bug in ASCII85Decode.decode() method and unify impl. between ver. 2 and 3.
The bug manifested solely in the Python 2 version of the code, which has been
deleted now (sorry, but it's a little cumbersome to maintain **two** distinct
implementations of an algorithm at once - the one stored now is efficient and
works seamlessly for both Python 2 and Python 3!).

TO-DO: the current decode() algorithm doesn't verify the ending `~>` EOD
sequence in an ASCII85-encoded sequence.