Rename private methods and attributes of `PdfFileReader` with a leading `_`.
This renaming was motivated by several methods (and variables) that performed
private actions but were shown as public by their nomenclature. The most
delicate refactoring here was `PdfFileReader.reader()` as, during the change,
there were lots of similar occurrences (of which only _one_ was in the end the
one looked for).

Lower camelCase changes were also added were needed.