Rename `pypdf4` module name to `pypdf`.
This change was also hinted to in [a PR#14
comment](https://github.com/claird/PyPDF4/pull/14#issuecomment-420702997). The
advantage here is that future versions of PyPDF will be able to rely on the
same module/package name, without having to change it. The name of the project
outside the code (in documentation, web pages etc.) shall however be maintained
as PyPDF4.