Perform HEAVY restructuring on how the XRef Table and XRef Streams are read in `PdfFileReader`.
The Cross-Reference Stream is read more than it was before and its data stored
in `PdfFileReader._xrefStm` variable. Additional work must be performed to
comply with the other ISO 32000 specifications.
A few other unit tests have been added and, as can be seen, all of the relevant
ones pass.

It is to note that the work initiated by this commit isn't completely over and
that, at the time of this writing, `PdfFileReader` doesn't fully conform to PDF
1.5+ specifications for reading the XRef Streams (see paragraph 7.5.8.4,
"Compatibility with Applications That Do Not Support Compressed Reference
Streams").