MAINT: PdfReaderProtocol (#1303)
PyPDF2 has some dependencies that make proper typing hard:

* PdfReader has the pages property which returns a List[PageObject]
* PageObject has the pdf property which returns the PdfReader it belongs to

The simplest solution would be to put both classes in the same file, but that makes PRs hard to read. Additionally, bigger files mean merge conflicts happen more often.

Another solution is to just not use type annotations for one of the objects (or use `Any` as the type)

The solution implemented in this PR is to define a `Protocol` (PEP 544): A protocol just states which methods a class is expected to have (with their function signature). It's duck typing: If it walks like a duck and it quacks like a duck, then it must be a duck.

So we define the expected behavior instead of referencing to the specific class.

typing.Iterable is an example for a Protocol. In the Java world, one would call this an interface.