tests: Use pytest.warns() for warnings, and .raises() for exceptions
Replace the warning-as-exception checks with use of `pytest.warns()`
that is more semantically correct and works correctly when the tests
are run without -Werror (e.g. because -Werror tends to cause test suites
to crash on irrelevant deprecation warnings from other components).
While at it, replace the homegrown exception checks in test_orientations
with `pytest.raises()`.