TST: Replace pytest.mark.external by enable_socket (#1657)
pypdf has three types of tests: (1) Tests that work only with the repository itself, (2) tests that require the sample-files git submodule, and (3) tests that require PDF documents from external websites.

As people or projects might not be able to execute (2), they are marked with `@pytest.mark.samples`

As people or projects might not be able to execute (3), they are marked with `@pytest.mark.enable_socket` (formerly @pytest.mark.external).

We regularly had issues with people forgetting to mark their tests with `external` (see #1632).

The [pytest-socket](https://pypi.org/project/pytest-socket/) pytest plugin disables the usage of sockets on the whole project, except for the tests marked with `enable_socket`. This was, we can guarantee that the CI will catch missing `@pytest.mark.enable_socket`.