TST: Use fixtures to create files in temporary folders (#1730)
This change should make tests easier to read and ensure that the cleanup-step cannot be forgotten.

If the test run fails, you can see where the file was written. The file is not directly deleted.

However, if the run does not fail (loudly) but you want to inspect the generated PDF, you have to 'assert False' or similar. That is  harder to figure out than commenting out a Path.unlink() line