ROB : Prevent loop in Cloning
fixes  #1767
the issue is due to object 589/0 : this object corresponds to 2 object the file trailer but also an XObject corresponding to the filled text ("test") because of that the object tries to be duplicated inducing the loop. This loop is not normal and I've fixed as a robustness improvement. If you run your code the text "test" will be missing as it is hidden by the trailer object.