ROB: Prevent loop in Cloning (#1770)
An issue was occurring with object 589/0 of file https://github.com/py-pdf/pypdf/files/11138472/test.pdf, which caused a loop during cloning due to its correspondence with both the file trailer and an XObject for filled text ("test"). This behavior was not intended, and a robustness improvement has been made to prevent the loop. Please note that if you run your code, the text "test" may be hidden by the trailer object.

Fixes  #1767