Add test for matrices given to visitor_text
Check the coordinates computed from the tm_matrix and cm_matrix
that are given to `visitor_text` when extracting text.
The test computes the coordinates of three letters in different
parts of a test page based on the matrices given to the `visitor_text`
function and checks, if they are roughly where they should be.