DEV: Move title check to own workflow (#2384)
Move the title check script introduced in #2378 into its own workflow where it's triggered by `[opened, reopened, edited, synchronize]`, where the difference is in the addition of `edited` type. This allows the workflow to be triggered on editing any part of the PR, most importantly the title attribute to fix it not following conventions. Otherwise, it would require the user to either push a new commit or open/close the PR which is not as intuitive to fix.