ENH: Add "layout" mode for text extraction (#2388)
The `PageObject.extract_text` got a new `extraction_mode` parameter. The old type of extraction is called "plain" which aims more at extracting text in a way that would be useful for NLP or a Text-to-Speech (TTS) system.

The new `extraction_mode="layout"` aims at visually representing the PDF. This is useful for detecting/extracting tables.