TST: Skip tests using fpdf2 if it's not installed
The Debian ecosystem installs only packages which are also present
in the Debian ecosystem for testing. This includes pytest, but it
does not include fpdf2.

Adding the line 'pytest.importorskip("fpdf")' within the test
and before the import ensures that pytest skips the test in
case fpdf is not installed.

Closes #2408