STY: consider images inside pages patterns.
Added code to detect patterns in "_get_ids_image".
To avoid any conflicts with images that could be located directly in a page or images using the same ID in differents patterns, images ids under patterns are returned in this form :
"/Pattern/patternNameHere/imageNameHere"

Added code to deal with Pattern images in "_get_image".