BUG: test_image_without_pillow cannot find pypdf
test_image_without_pillow runs a generated script which causes the
Python path to exclude the current directory.  The generated script
tries to import pypdf and either cannot find it or it finds the
version in pyenv instead of the version being tested.  Add "." to
PYTHONPATH so the correct version of pypdf is used.

Closes #2849