TST: Add LzwCodec for encoding (#2883)
This PR adds an abstract Codec class as well as an LzwCodec encoder implementation.

While we might only need decoding for pypdf, having both decoding and encoding in one class massively helps with testing. We can still get it wrong, but it's harder to get both the encoder and the decoder wrong in a consistent way.