_writer: Implement flattening
This is a vibe-coded attempt at implementing flattening of pdf forms. It is
currently able to flatten text fields and button fields.

It is not at all integrated with the rest of the pypdf code. It introduces
a dependency on pdfminer.six to calculate font widths. This should also be
possible with some pypdf code but I don't think that pypdf has correct font
metrics. (I tried with STANDARD_WIDTHS from
pypdf/_text_extraction/_layout_mode/_font.py , but text started overflowing
certain fields.

So, basically a proof of principle.