ENH: PdfWriter: Add method that adds appearance stream as XObject
This patch adds the add_apstream_object method, which takes a
page, an appearance stream, an object name, x and y offsets and
optionally a font resource. It turns these into an XObject and
then merges the XObject drawing commands to the page.

This code can be used to flatten an annotation.