ENH: PdfWriter: Implement flattening of annotations
This patch implements flattening of annotations by adding the
bool "flatten" to update_page_form_fields. It defaults to
False. When set to True, it will add the annotation's
appearance stream to the page as an XObject. Note that this
will not remove the annotation itself. This, however, can
be achieved, for instance, by calling
remove_annotations(subtypes=None) after having flattened
all annotations.

When calling update_page_form_fields with the values of
the fields dict set to None, it will not change existing
values and instead only flatten what value was set
previously. This can be used to flatten an existing form
that has already been filled out. At a later point, this
might be useful for adding a specific flatten method.

Note that this patch does not change the text appearance
stream at all, which means that it does not automatically
change font size, it does not wrap text, nor does it try to
fit text inside a box.