ROB: ignore faulty trailing newline during RLE decoding (#3355)
Found PDFs from Dalim software with multi-encoded streams: inner stream is RLE, outer stream is FLATE. The inner stream contains a trailing newline char that breaks the RLE decoding. It seems that there was in some Dalim version a systematíc error that included the bytes of the inner stream just from raw PDF bytes with the trailing newline before "endstream".
This is fixed with the changes by ignoring the trailing newline and raising a warning instead of an exception.