ROB: Switch Font class to FONT_METRICS
The _font_widths.py file contains a set of standard
widths for the core 14 Type 1 PDF fonts. However,
it is incomplete; for instance, it sets all Helvetica
variants to the same set of widths. Since we have
imported FONT_METRICS from the original Adobe core
fot afm files now, which are complete, complete, port
over the Font class to use that data.