MAINT: Remove _font_widths.py
This file contained STANDARD_WIDTHS, which was contained
font widths for the core 14 Type 1 PDF fonts. This is no
longer used since we imported FONT_METRICS from the afm
files.

Also, it would appear that the _font_widths.py did
not include the associated licensing information,
which, with its removal, is now also remedied.