ENH: _codecs: Add 14 Adobe core font metrics
This patch adds a new file with the font metrics
for the 14 core Type 1 pdf fonts. The file was
inspired by the pdfminer.six project, where a
very similar one is called fontmetrics.py.
The information itself is generated by a
separate file added with this patch:
resources/get_core_fontmetrics.py.

The PDF specification expects a pdf reader to
include these font metrics. For the purposes
of pypdf, these metrics are needed to build
build_font_width_map font widths when a font is
not embedded.