ENH: _font: Init FontDescriptor from font resource
This patch adds a class method to initialise FontDescriptor
using a font resource dictionary. For now, this returns
either a font name and FontDescriptor defaults, or, in case
of one of the 14 core fonts, the associated font metrics.

For future development, this class can be extended with
code that collects character widths, which is now present
in multple places, such as:
 - pypdf/_text_extraction/_layout_mode/_font.py
 - pypdf/_cmap.py

... and code that collects font descriptor information,
which is now present nowhere yet, but would be useful
for generating text streams.