ENH: _font: Set some defaults for the FontDescriptor class
This patch introduces some defaults for the FontDescriptor. This
might be useful if we would like to migrate existing code for
collecting font widths to the class, while not having all code
in place to also collect all associated other information about
the font.

The defaults I took from Google Gemini...