MAINT: Refactor code from _writer to _appearance_stream
This patch introduces a new module - appearance_stream - and
copies two methods from _writer to this new module. Currently,
these methods are needed to develop an appearance stream for a
text annotation. They are:
update_field_annotation (renamed from _update_field_annotation)
generate_appearance_stream

The update_field_annotation was a PdfWriter method, which means
that the current code needs some refactoring, since it now has a
circular import of PdfWriter.

Other than changing self to writer in update_field_annotation,
and changing the code in PdfWriter to call update_field_annotation
from _appearance_stream, and it avoids a little bit of casting in
one place. For the rest, this patch changes nothing.

In a future change, we might want to make a class
TextAppearanceStream based on generate_appearance_stream, with
.from_annotation(Annotation) as a class method (based on
update_field_annotaion). scale_text would also be a method in
this class.