MAINT: _appearance_stream: Make variables more clear and readable
Code in _appearance_stream used various rather cryptic variable names
that, for some coders, made it hard to understand what the code was
doing. This patch tries to clarify those variable names to make it
easier to understand what's going on, and make it easier later on
to add functionality.