ENH: TextAppearanceStream: Add method to scale and wrap text
This patch adds a method to scale and wrap text,
depending on whether or not text is allowed to be
wrapped.

It takes a couple of arguments, including the text
string itself, field width and height, font size,
a FontDescriptor with character widths, and a bool
specifying whether or not text is allowed to wrap.

Returns the text in in the form of list of tuples,
each tuple containing the length of a line and its
contents, and the font size for these lines and
lengths.