MAINT: _appearance_stream: Rename font_height to font_size
This patch removes the variable name "font_height", because it
means the same thing as font size. I think that font_height was
introduced previously to distinguish between a font size found
in an annotation's default appearance and the size set by a
user.

To be consistent, also use the variable user_font_name when it
pertains to a user choice, and font_name for a font name found
in a default appearance.