DEV: Use Mapping instead of dict in type hint of update_page_form_field_values (#3490)
Having the parameter typed as a dictionary does not allow for stricter
type hints on the user side, like `dict[str, str]`, as a dictionary is
considered mutable. Typing our method as `Mapping` indicates that we
do not change the corresponding values inside it and does allow for
stricter typing in user code.

Closes #3453.