MAINT: _cmap.py: Derive default width from CORE_FONT_METRICS
Default width, if not given or defined in a font resource, is
calculated either by averaging all font widths or, in the case
of unembedded fonts, calculated as the width of two spaces.
Previously, this width was part of a separate table in
_cmap.py. This patch changes the associated logic and instead
reads space width from CORE_FONT_METRICS, if a default width
is not given or found elsewhere.

Note that the patch removes the info for four space widths from
pypdf, i.e., the four Helvetica-Narrow fonts. However, these
are not among the fourteen Adobe standard fonts, so, they
should be entirely embedded (and have a default value or be
able to compute one by averaging all font widths).

Note that it's slightly inconsistent now, that CID fonts that
don't have a default width defined will need to get it from
the value set by the user or when, by accident, the have the
same name as one of the fourteen Adobe standard fonts, whereas
other embedded fonts get a default width equal to the average
character width, and the fourteen standard fonts get a default
of two space widths. However, this inconsistency was already
part of _cmap.py before.