MAINT: Fix tests after changing font and margin calculations
This patch fixes four tests. The fixes for test_appearance_stream
are more or less self-evident, since the associated streams have
changed.

In test_writer.py, I changed test_da_missing_in_annot, which
previously checked that "0 Tf" would not be in the appearance
stream, indicating that font had been resized. Unfortunately,
the changes in margin and font size calculation resulted in a
font size of 9.0, causing "0 Tf" to be matched and the test to
fail. Check for " 0 Tf" instead.

The test_field_box_upside_down test tested a specific
appearance stream. Changed in accordance with the changes in
margin calculation.