MAINT: Update filelock version to 3.20.1
This closes a GitHub dependency alert.

Please note that this update is not compatible
with Python 3.9 anymore. If you still develop on
Python 3.9, consider using the `dev.in` file instead.
It should still be compatible as far as I am aware,
but please keep in mind that it might expose you
to CVE-2025-68146.