MAINT: Layout mode text extraction: Port to new Font class
This patch ports the layout mode text extraction code to the new font class.

This introduces one test failure, which itself appears to derive from a
misconception about space width in the original Font class.

Previously, a layout mode font was initialized in _page.py as follows:
  fonts[font_name] = _layout_mode.Font(*cmap, font_dict)

*cmap, in this case was the return value of build_char_map, which consists of:
  - Font sub-type;
  - Space_width criteria (50% of width);
  - Encoding map;
  - Character-map; and
  - Font-dictionary

Notice that build_char_map does _not_ return the width of a space, but the
width of _half_ a space. However, if we look at the arguments to the layout
mode Font class, clearly the class expects to be passed the full width of a
space. This is also clear from the word_width method in the layout mode Font
class, which substitutes a missing width with 2 * space_width. It follows
that the layout mode Font class _expected_ to be passed a full space width,
but really was only passed the width of half a space.

When porting to the new Font class, this becomes problematic when calculating
text width, because the new Font class uses self.character_widths["default"]
as a fallback for a missing width, which is approximately (and in many cases
exactly) the width of two spaces. This in turn causes problems with text
extraction in cases where the width of a space itself is missing ("The Crazy
Ones"), and cases where a font with a missing character width is calculated
wider than before.

For the first issue, this patch introduces a work-around that also exists in
the conventional text extraction code, that is, dealing with missing space
width separately.

For the second issue - this causes one test to fail:
the test_layout_mode_text_state in tests/test_text_extraction.py. This is
entirely due to the existence of a unicode private range character in the
file.