MAINT: _page.py: Add asdict for json dumps of fonts
The layout-mode Font had a to_dict method that returned a json
object of a Font object. Instead if directly copying that method,
use asdict() from dataclasses in _page.py and use asdict in the
associated test in test_text_extraction.

I DID NOT TEST THE DEBUG CODE IN PAGE.PY MYSELF.