DEP: Block common page content modifications when assigned to reader
Closes #2260.

This would previously lead to pages being written uncompressed, although
the corresponding dictionary header declaring the filter to be
FlateDecode. As a PdfReader is considered to be read-only, this change
seems like the most suitable one for fixing this.

With this change, it might be required to change own code previously
relying on the more or less broken functionality, which is especially
bad for shadow processing.

Possible approaches to fix user code:

  * Use `PdfWriter(clone_from=...)` to add all pages to the writer.
  * Use the return value of `writer.add_page(page_from_reader)` which
    correctly belongs to the writer to apply possible modifications
    after adding the page itself (instead of before adding it).