MAINT: Fix test output after fixing character widths
Previously, character widths were not computed for type1 and
TrueType fonts when encoding was a string. For one test, that is,
test_text_extraction_layout_mode in tests/test_workflows.py, this
meant that all character widths were treated as one space width.

Now, the real widths are used, which changes the output of the
test significantly, but in keeping with the intended output. This
patch implements the new output.

To be sure, I counted the number of newlines and words in both
versions, and they are exactly the same, so no spaces were
accidentally omitted between words in the new version, nor were
they added, since the new version has fewer spaces than the old
one.